runs separated integration tests in parallel
