+++ integration/test/scaling/project.yaml
@@ -1,0 +1,23 @@
+version: 1
+
+test:
+  env:
+    - "VAULT_TOKEN=mytoken"
+    - "CIRCLE_SHA1=${CIRCLE_SHA1}"
+    - "CIRCLECI=${CIRCLECI}"
+    - "CLUSTER_NAME=${CLUSTER_NAME}"
+    - "COMMON_DOMAIN_GUEST=${COMMON_DOMAIN_GUEST}"
+    - "GUEST_UPDATE_ENABLED=true"
+    - "TESTED_VERSION=${TESTED_VERSION}"
+    - "REGISTRY_PULL_SECRET=${REGISTRY_PULL_SECRET}"
+    - "IDRSA_PUB=${IDRSA_PUB}"
+    - "GUEST_AWS_ACCESS_KEY_ID=${GUEST_AWS_ACCESS_KEY_ID}"
+    - "GUEST_AWS_SECRET_ACCESS_KEY=${GUEST_AWS_SECRET_ACCESS_KEY}"
+    - "GUEST_AWS_SESSION_TOKEN=${GUEST_AWS_SESSION_TOKEN}"
+    - "HOST_AWS_ACCESS_KEY_ID=${HOST_AWS_ACCESS_KEY_ID}"
+    - "HOST_AWS_SECRET_ACCESS_KEY=${HOST_AWS_SECRET_ACCESS_KEY}"
+    - "HOST_AWS_SESSION_TOKEN=${HOST_AWS_SESSION_TOKEN}"
+    - "AWS_REGION=${AWS_REGION}"
+    - "AWS_API_HOSTED_ZONE_GUEST=${AWS_API_HOSTED_ZONE_GUEST}"
+    - "AWS_INGRESS_HOSTED_ZONE_GUEST=${AWS_INGRESS_HOSTED_ZONE_GUEST}"
+    - "KEEP_RESOURCES=${KEEP_RESOURCES}"
